pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vn-legal-rag"
version = "1.0.0"
description = "3-Tier RAG System for Vietnamese Legal Documents with Knowledge Graph Integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Vietnamese Legal RAG Team", email = "contact@example.com"}
]
maintainers = [
    {name = "Vietnamese Legal RAG Team", email = "contact@example.com"}
]
keywords = [
    "rag",
    "retrieval-augmented-generation",
    "legal",
    "vietnamese",
    "nlp",
    "knowledge-graph",
    "graphrag",
    "legal-tech",
    "question-answering",
    "information-retrieval"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # LLM Providers
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",

    # Database & Storage
    "sqlalchemy>=2.0.0",

    # Text Processing & Similarity
    "jellyfish>=1.0.0",  # Jaro-Winkler similarity for Vietnamese text

    # Scientific Computing
    "numpy>=1.24.0",
    "scipy>=1.10.0",

    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "tqdm>=4.65.0",  # Progress bars
]

[project.optional-dependencies]
# Core embeddings (required for semantic search)
embeddings = [
    "sentence-transformers>=2.2.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
]

# FAISS for fast vector search
faiss = [
    "faiss-cpu>=1.7.4",
]

# Full installation with all features
full = [
    "vn-legal-rag[embeddings,faiss,dev]",
]

# Development dependencies
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",

    # Code Quality
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0",
    "isort>=5.12.0",

    # Type Stubs
    "types-PyYAML",
    "types-tqdm",

    # Documentation
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",

    # Jupyter
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

# Evaluation & benchmarking
eval = [
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/vn_legal_rag"
Documentation = "https://vn-legal-rag.readthedocs.io"
Repository = "https://github.com/yourusername/vn_legal_rag"
"Bug Tracker" = "https://github.com/yourusername/vn_legal_rag/issues"
Changelog = "https://github.com/yourusername/vn_legal_rag/blob/main/CHANGELOG.md"

[project.scripts]
vn-legal-rag = "vn_legal_rag.cli:main"
vn-legal-query = "vn_legal_rag.cli:query_command"
vn-legal-eval = "vn_legal_rag.cli:evaluate_command"

[tool.setuptools.packages.find]
where = ["."]
include = ["vn_legal_rag*"]
exclude = ["tests*", "docs*", "scripts*", "data*", "plans*"]

[tool.setuptools.package-data]
vn_legal_rag = [
    "py.typed",
    "config/**/*.yaml",
    "config/**/*.yml",
]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | data
)/
'''

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "data",
    "__pycache__",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["vn_legal_rag"]

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradually enable
ignore_missing_imports = true
exclude = [
    "tests/",
    "scripts/",
    "data/",
]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=vn_legal_rag",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["vn_legal_rag"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
